Change font size of division by 0 message
Division of 0 error message didn't fit on one line.

In html: add display-normal class to display div.

In css:
Move font-size from display id to display-normal class.
Add display-error-msg class with smaller font size declaration
to fit on one line.
With font-size in display id it will always have greater specificity
than a class. With both font-sizes in classes, the lower declaration
will have higher precedence.

In JavaScript: add if statement in updateDisplay to add the
display-error-msg class to the display div if the result to display
is the division by 0 error message.